refactor: move config to src/config/app-config
