+++ src/main.ts
@@ -1,6 +1,6 @@
 import { createAppAuth } from "@octokit/auth-app";
 import { Octokit } from "@octokit/rest";
-import { loadConfig } from "./config";
+import { loadConfig } from "./config/app-config";
 import { createLogger } from "./infra/logger";
 import { RealCoderClient } from "./coder-client";
 import { WebhookRouter } from "./webhook-router";
